Storage policy cleaned, storage class explicit loading
Now able to load a storage class explicitely, out of any storage policy context. This is necessary for the metacd-http to process load-balancing requests with stgclass constraints. In facts, the metacd_http simply wraps the calls to metautils/lib/lb.h API.

A corner case has been managed, when a storage class exists but must not have fallbacks. It was necessary to declare the storage class with an empty configuration.

Then, the code has been cleaned, simplified and reordered by purpose.

Change-Id: Ibd66594d53666f5ca4d1fab6d4226e1dc17ba95d