sqlx: fixed db handle init when used in post_config
In meta0, a base was opened *before* the all the iserver's structures were
ready and bond together. That base was initiated with no known election manager,
then returned in the cache, and reused later as is. But for replication to take
place, such an information is mandatory (thus checked with an assert()).