sqliterepo/cache: prevents threads starvation
Get rid of the old model with a fixed-number of GCond used to bucketize
all the DB accesses.

Now one couple of GCond (normal, prio) is associated to each cache item,
and this is better since the lifetime of each item is the whole
execution.

In addition, the notion of maximum bases used has been reviewed to allow
slight changes on the effective maximum. Now we propose one "hard"
maximum used to allocate structures at init time, and one "soft" maximum
that can be freely changed between (0,hard].