sqliterepo/election: cleaner trasitions the FSM
The goal of the commit is to prevent entering a state with several
pending callbacks that can trouble the transitions and cause messages
storms.

Only the completion events make the FSM (finite state machine) step
forward. All the "interruptions" events are delayed until the normal
actions are all done.

Summary of the changes:

* general renaming of the states to more intuitive names that denote the
  action that is pending. LEADER -> MASTER, LOST -> SLAVE.
* new states added, so that at most one action is pending in each state.
  specially PRELOST that becomes [ASKING,CHECKING_MASTER] and CANDOK
  that becomes [WATCHING,LISTING].
* new state added to leave-and-then-fail, instead of leaving then
  retrying.
* Tests, tests, test. Despite the FSM is not 100% covered yet, the major
  transitions are now checked.
* Timers simplified: since we wait for actionns before leaving an FSM
  state,.the necessity to retry an election is vastly simplified and
  only occurs from the NONE state.
* A jitter has been introduced before sending a ping (USE) from a final
  state.

Despite the result should be more stable, robust and resilient, this is
a big set of changes and one shoould take care of possible unexpected
effects. Especially of elections that could go more often than necessary
in the FAILED state, and then undergo a 5s timer before the retry.
(Prior to this commit, we would have loop-retry more often than
necessary.)