proxy: Add bulk delete content
Permit to delete many content in one go for one container.
Modify the format of cli command 'openio object delete' to return a List.Lister.
Modify tests that depended on the old format response to use the new one.

The format of a request to the proxy is a json '{"contents":[{"name":"..."},{"name":"..."}]}'
and the url is /content/delete_many.
The format of the answer is a json '{"contents":[{"name":"...", "status":"...", "message":"..."}]}'