sqliterepo: Prevent a potential deadlock while exiting
During the cleanup routine of an election manager, the lock is acquired
to prevent the elections to be modified. But at that time, the elections
have left and the worker threads exited. The lock is not mandatory.

Unfortunately, the queue of pending notifications is maybe not empty,
and the notification hook still may want to acquire the lock. so let's
just ignore that critical section.

(cherry picked from commit 2b2929f7d74eb4820baeb14f7858e9a919f6c530)