New LB tests for tight platforms
Check how many unbalanced repartitions we get on platforms where
there is not enough locations to ensure a perfect data securing.

For example, when there is 3 host with 4 services each, and we ask
for 9 services, the load-balancer sometimes returns 4 services from
the first host, 2 from the second and 3 from the third,
whereas we expect 3, 3 and 3.