oio: support restoration of container by range
When a dump operation is launched, the value
of X-Manifest-Container header must be sent when using
restoration operations using range.

This header can also be retrieved by a HEAD on container.

Please note that restoration by range must be a multiple of 1 MB,
as empty block may be inserted to avoid header split.

When untaring a dump on Linux filesystem, you should provide option
--ignore-zeroes to avoid 'A lone zero block at %d' error.