Sqliterepo: Fix replication error on admin table
There was a problem of desynchronization of the rowid value.
It is use internally by sqlite as the true primary key of the table.
We used this value as the primary key to have exactly the same
database, but it is not necessary for bases to have the same rowid for
the same information.