Rdir assignment to META2 on volume bootstrap
The same facilities used to assign Rdir services to rawx services
are reused in this patch to assign some to META2 services.

For the time being, the bootstrapping is done when the initial
volume bootstrapping is done.

A future patch will bring a new command to do that separatly from
the volume bootstrap.