sqliterepo: directly dump database file, do not load more than 1GiB
Before this commit, a local copy of the database file was created, and
the data was sent to the network from the copy. Now, the data is read
directly from the database file.

This commit also sets a maximum on the size of database dumps, in order
to avoid memory exhaustions. This maximum is 1GiB by default, but can be
set to 0 to completely disable the DB_RESTORE mechanism, or set up to
2047MiB, with the parameter "sqliterepo.dump.max_size".