Python API: forward timeouts, compute a deadline
The 'patch_kwargs' functionality was used only on a few ObjectStorageApi
methods, resulting in the read_timeout not being applied on all requests
to oio-proxy. This is now fixed.

Also, if a timeout is provided, a deadline is now computed for internal
usage, and a timeout is computed accordingly for each subrequest.