sqliterepo: reset master elections on ZOO_SESSION_EVENT
A long time ago, we used to reset all elections on such Zookeeper
event. We stopped doing this when we started to use several
connections to Zookeeper, because it seemed a bad idea to reset
everything on a single connection failure.
Unfortunately, this leads to elections staying forever in the same state
while their Zookeeper node has vanished.

Now, after getting an unbound (no path is provided) ZOO_SESSION_EVENT,
we will reset all elections in the master state, to avoid any infamous
"double master" situations.