Python API: report EC chunks errors only once
Because of a recent modification, some errors were reported twice (or
more!), and the Python API attempted to commit the same chunk several
times. This was fortunately blocked by the meta2
("Content broken and irreparable (HTTP 500) (STATUS 425)"), and the
whole object upload was considered a failure, even though the chunks
quorum was reached. This is now fixed.