Allow lifecycle to work with sharded buckets
For S3 buckets which have been sharded using "container_hierarchy" swift
middleware, allow to recurse on all shards when applying lifecycle
rules.

This behavior is enabled with a command line argument.
The set of containers to recurse on is established by doing a prefixed
listing (the prefix being the bucket name and the '%2F' delimiter).

The metadata of listed objects will be enriched with
- the name of the actual container they are stored in (under the key
'container');
- the name they have been given in this container (under the key
'orig_name').

All tests against filters are unchanged, but API calls will use
`orig_name` and `container` instead of object and bucket names.