Account: Fix account refresh on missing values
There was an error when damaged_objects or missing_chunks were missing.
This happened when a container is created before oio-sds 18.10 and no object
is created after an update > 18.10.
The fix set the value to 0 but if some object had missing chunks before it will not
be printed and therefore the value would be erroneous.