Python API: restore sys.stderr in case of log error
When logging to /dev/log and running with Python 3, if /dev/log becomes
unavailable, the logging system writes an error to sys.stderr. But since
we replace sys.stderr by a logger, we end up with an infinite recursion.

Now, when detecting a logging error, we will restore the real stderr.