proxy: Allow to force the request on master service
It was already possible to do this by adding 'proxy.force.master'
in the proxy configuration. But this was applied to all requests.

If we wanted to force the master only on a specific request,
we can add the 'x-oio-force-master' header in the request.
But this header was only used to list the objects in a container.

Now, it is possible to specifically force the master on all meta2 requests.