Python API: accept a scheme in oioproxy URL
Refactor the code so we handle oioproxy's URL the same way, whether it
comes from the application's configuration object, a keyword parameter
or the namespace configuration file.

If there is no scheme in the URL, consider it is HTTP. If there is a
scheme, use it (but only HTTP will work at the moment).