conscience: refuse services from other namespaces
To mitigate the effects of configuration errors which lead to services
registering to the wrong namespace, check it while receiving service
lists.

There is still a possibility to register wrong services: not
providing any namespace name. At the moment, this is done only by the
"cluster wait" CLI command, which gets its list of service from the
conscience (the services have been registered before).