proxy: return code 503 a storage policy is not satisfiable
The fact that a storage (or service) policy is not satisfiable comes
more often from a lack of availables services than from a
misconfiguration, therefore is it more logical to return a "retry later"
status instead of an "internal server error".

Also return a code 400 when receiving an unknown policy name.