meta2: allow container deletion even when there are ghost chunks
If an object is rebuilt at the same time as it is deleted, the M2_RAWADD
operation can insert chunk entries not attached to any "content". This
used to cause problems when trying to delete the container: it was
reported as not empty. Now we will mainly check the "contents" table.