event-agent: apply policy regex on events before forwarding
The configuration of the "notify" event-agent filter now allow to set
regexes for object storage policies. The event will be forwarded to a
tube whose regex matches. Regexes are matched in alphabetical order and
processing will stop at the first match. If there is no match, the event
will be forwarded to the default tube.