Make sure selected spare chunks while moving data respect fair location constraints
The max number of chunks on a host was based on strict location constraints thus when moving chunk some irrelevant location were selected. This commit changes the rule, now when moving data we calculate the max of chunks on a host using fair location constraints to make sure the selected candidates are relevant.