events: improve rabbit reconnection
If a rabbit was DOWN, "_q_reconnect" was not detecting anything so the
socket was still used (and the connection to the rabbit was never able
to work again).
Now, the rabbit is disconnected when stats are not available.
With this commit, the meta2 is able to deal with rabbit failures.