proxy: check if delete operation did actually delete
Sometimes it will create a delete marker instead of deleting an object
version. In that case we must return the version of the created delete
marker.

Notice that the return value of ObjectStorageApi.object_delete() changed
type: it is now a tuple with a boolean telling if a delete marker has
been created of deleted, and the version of the object (or delete
marker) which has been deleted (or created).