crawler: add new rawx crawler that cleans up the orphan chunks
Cleanuporphaned crawler cleans up the chunk identified as orphaned.

The crawler browse orphans folder which contains symlinks to orphan chunks.
In the process a call to meta2 master db is made to make sure that the
chunk is actually an orphan chunk.

Additionnaly to this, we wait a certain time (configurable) before
deleting an orphan chunk. It is an extra safety measure because we
are dealing with data deletion.