meta2: reallow draining root container
Draining a root container is triggered by the meta2-crawler when all
shards are fully drained.
It is the way to update the draining status in the root which is used by
PCAv2 to know if a bucket is drained or not.
If we block draining on the root container, it will remains in
draining=Needed forever.