Optimize the command used to list ongoing mpu
To list the unfinished MPUs we read all objects
in segments container and then filtered to find all markers.
Following this change, instead of listing all objects in the
segment container, we will list only the markers present in
the container. This change is made  by adding a filter to
the SQL query directly.