sharding: Fix clean timeout
Now that the deadline is monitored during replicated cleaning,
it often happens that this delay is slightly exceeded
(due to somewhat long operations).

To prevent the cleaning fails during sharding, we need
to have as few errors as possible on these operations
(including timeouts).
By keeping a long timeout on the client side and putting
a short timeout on the server side, the risk of ending up
with a timeout is very reduced.