encryption tool: encrypt user metadata
Encrypts each item of custom user metadata using the object key provided
by the kms and an IV that is randomly chosen for that metadata item.
Associated crypto-metadata is appended to the encrypted value:
  X-Object-Transient-Sysmeta-Crypto-Meta-Private1:
    E(value1, object_key, header_iv_1); swift_meta={"iv": header_iv_1,
                                                    "cipher": "AES_CTR_256"}
Store a single copy of the crypto meta items that are common to all
encrypted user metadata:
  X-Object-Transient-Sysmeta-Crypto-Meta
    cipher=AES_CTR_256"; key_id={"v": "1", "path": path, "sses3": True}
Remove the unencrypted custom user metadata.