sharding: Trigger a VACUUM even when local cleaning timeout
A local cleanup has a deadline to delete as many entries as possible.
It is possible that this deadline is very slightly exceeded
(the meta2 service detects the end of the deadline with a slight delay).
In this type of situation, the VACUUM must still be triggered
because the database was cleaned but did not respond in time.