python: stop writing logs to stdout on log level lower to INFO
python program with log level lower to INFO automatically sends logs
to stdout. While this can be useful for development it cause double
logging when syslog is configured and when the software is run as a
service on production servers.

To force writing logs to stdout while with DEBUG log level, just set
verbose.