encryption tool: find right secret for object list
Add script that reads extracted secret file and JSON file containing all
objects for a specific bucket sorted by version.
Try to find bucket secret associated to each object and add it in the
object dict from JSON file.
If object is empty, it is not rewritten.
If right secret is not found for one object, the script exit code is 1.