CLI: Make container checking fast again
With a value of 2, container checking is extremely slow
(3 hours compared to more than 4 days for the same container).
There are currently scripts that use this command in a loop.
Stopping these scripts would require starting all over again.
To avoid this, we need change the default value so that the
next command is no longer launched with the value 2.