conscience: add fallback value in accumulator (OBSTO-3789)
* current accumulator in expr: `stat.key`. if key is not found, an error
  is returned and the scores is set to -1
* new accumulator with fallback in expr: `stat.key:"xxx"`. if key is
  found, key is returned else xxx is returned

the test remains TestConscienceFunctional.test_get_score which ensure
right scores. `* (num stat.unknown_stat:"1")` has been happened to the
score_expr of the echo service in oio-bootstrap.py in order to make the
test_get_score test unchanged and validated at the same time that the
fallback syntax for accumulator is working.