tests: do not always call _reload_meta
This call dates back from a time when meta2 services needed to know the
list of rawx services. Service selection has been done by oio-proxy for
a while now, we don't need to force meta2 services to reload the list.

Also, when reloading the list, do not fail on error, just print a
warning.