proxy: do not retry after timeout right after deadline
When the deadline is close, we set short connection and read timeouts.
Running into these timeout does not mean the remote service is
unavailable, it just needs a little more time to connect or handle the
request. Retrying the request with an even shorter timeout does not make
sense, we'd better return "Deadline reached" immediately.

We witness this scenario especially with long listing requests,
which struggle to respond within the deadline.