Redirect to the shard containing the marker when there is a version marker
Previously, if the upper was present in a shard with multiple versions,
a listing containing a marker for one of its versions would ignore
the others and proceed directly to the next shard.

Now, all versions are correctly listed, even when the following object
corresponds to the upper with only the first Unicode character as a suffix.