event-agent: set a long retry delay after IO errors, allow no retry
No need to retry immediately after an IO error. We don't know when the
drive will be replaced, but there is very little chance it will be
replaced in the following minute.

Also, we can now choose to no retry IO errors, and consider the chunk
is. This prevents from sending too many events to the deadletter topic
(we acknowledge them even if there were errors).