account-cleaner: Reduce the log level when a bucket has no container
This can happen when the container creation event was processed
before the bucket was created.
In all cases, the object and byte counts are correct;
the bucket was simply created (without an object upload).
The container will only be counted in the bucket
when there is a write to this bucket.