Prevent VACUUM error from causing the new shard creation to fail
The creation of a new shard may be preceded by a cleanup of entries
outside the range of the local shard copy.
At this stage, the CID of the new shard is not yet known because
we are only manipulating a copy of the parent database.

Additionally, the log was not using the correct CID compared
to the one used in the operation.