metautils: set a minimal timeout on every metadata request
Due to the fact that the timeout is derived from the request deadline,
some timeouts were set to a really low value (1µs). Now consider
"gridd.timeout.margin" as the minimal timeout we can set on metadata
requests. This prevents high-level services from timing out when talking
to lower-level services which may not respect the deadline.