python: Always use the chunk to rebuild as a last resort
For objects using replication, the chunk to be rebuilt was used
as a last resort only if rebuilding was allowed on the same rawx.

The use of the chunk to be rebuilt as a last resort no longer
depends on any option. It is systematically used if other copies
did not allow the chunk to be rebuilt (elsewhere).