meta2: Fix chunk deletion and replacement when chunk ID is stored
Depending on the configuration, chunks can be stored
in the meta2 database in 2 formats:
- the full URL of the chunk with its ID
- only the service ID of the chunk's rawx

Depending on when the format parameter was changed (or due to a bug),
a meta2 database can contain both formats.
Therefore, chunks searches in this meta2 database must be handled by
testing both formats (giving preference to the expected format).

Co-Authored-By: Aymeric Ducroquetz <aymeric.ducroquetz@ovhcloud.com>