proxy: Increase "proxy.bulk.max.delete_many" default value
The limit of 100 is too low and makes flushing slow (by default).
Also, the "openio container flush" command didn't control the limit value,
so the oioproxy service forced it to lower this limit for each request.

By increasing the value to 250, there will be fewer listing and flush requests.
Also, the "openio container flush" command now ensures that it no longer exceeds
the limit accepted by the oioproxy service.