sqliterepo: close DB before unlink
In order to avoid race conditions, we used to unlink the path to a
database before closing its handle. Recent versions of sqlite complain
about that ("sqlite3: (28) file unlinked while open"). There is very
little chance a race condition happens, because the closing code is
called while the thread has lock on the database cache. We can safely
close the database handle before unlinking the path.