event-agent: Fix storage policy selection during lifecycle transition
The `read_storage_mappings` method already sorts storage policies
in descending order of size.
Therefore, reversing this list was equivalent to starting with the
smallest size.
So the first element of the loop always matched because the size
associated with that storage policy is always less than or equal to 0.
Additionally, when negative sizes are configured (to continue mapping
old storage policies), a storage policy that should no longer be
associated with new objects was used.

Now, the loop will then start with the largest size until it finds
the appropriate storage policy.