change-storage-policy: Also check objects in transition
Unless the account is refreshed, storage policies no longer used
in the account remain visible in the stats (with a counter at 0).
But since transitions immediately change the storage policy stats
before modifying the data, it is possible to have a transition
in progress (potentially blocked) while the counter indicates 0.
So also check the containers of the accounts with a counter of 0
to make sure that there are no more transitions in progress.