sqliterepo: auto-fix partial DB_VACUUM
When vacuuming a huge database, it happens that none of the slaves
actually replicate the operation. In that case we are left with the
master ahead of the slaves with just one version. When detecting that
case, try to send a full copy of the database to the slaves (again).