sqliterepo: deal with slave timeout + rollback
When a master sees timeouts on all slaves, it rolls back the
transaction. But it happens that all slaves actually apply the changes
after the timeout. In that case we are left with all slaves ahead of the
master with just one version. When detecting that case, try to send a
full copy of the database to the slaves.

Co-Authored-By: Aymeric Ducroquetz <aymeric.ducroquetz@ovhcloud.com>