sharding: Manage excessively long pre-cleaning processes
Sharding operations create (by default) 50% empty space
on the new shards before they are vacuumed.
This vacuuming operation on databases larger than 1GB
can quickly exceed the previously allotted 30 seconds.
And if this operation exceeds 50 seconds (30 seconds
+ 20 seconds for attempting to open the database),
the actual creation of the new shard may fail because
the database is already occupied.

Now, a time limit of 60 seconds is given for completing
this vacuuming operation.
And the actual creation of the shard is retried if the
previous attempt failed to open the database.